backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
python-dotenv>=1.0
snowflake-connector-python>=3.0
requests>=2.28
slowapi>=0.1.8
sentry-sdk>=1.39
PyJWT>=2.8
pytest>=7.0
httpx>=0.24
pydantic-core
pydantic[email]
numpy
openai>=1.0.0
aiohttp>=3.8.0
pydantic-settings
python-multipart>=0.0.5
